Coding Standards: Cast `base_convert()` to an integer before arithmetic operations.
Follow-up to [35477], [53104], [56101].

Props justlevine.
See #64238.
Built from https://develop.svn.wordpress.org/trunk@61247


git-svn-id: http://core.svn.wordpress.org/trunk@60559 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/plugin.php
@@ -1427,7 +1427,7 @@ function add_menu_page( $page_title, $menu_title, $capability, $menu_slug, $call
 	if ( null === $position || ! is_numeric( $position ) ) {
 		$menu[] = $new_menu;
 	} elseif ( isset( $menu[ (string) $position ] ) ) {
-		$collision_avoider = base_convert( substr( md5( $menu_slug . $menu_title ), -4 ), 16, 10 ) * 0.00001;
+		$collision_avoider = (int) base_convert( substr( md5( $menu_slug . $menu_title ), -4 ), 16, 10 ) * 0.00001;
 		$position          = (string) ( $position + $collision_avoider );
 		$menu[ $position ] = $new_menu;
 	} else {
